fix: refactorig pagination
