+++ stolen-bikes-index-app/package.json
@@ -10,6 +10,7 @@
         "@types/node": "^12.0.0",
         "@types/react": "^17.0.0",
         "@types/react-dom": "^17.0.0",
+        "@types/react-router": "^5.1.16",
         "axios": "^0.21.1",
         "react": "^17.0.2",
         "react-dom": "^17.0.2",
