+++ stolen-bikes-index-app/src/components/DisplayCards/DisplayCards.tsx
@@ -1,0 +1,30 @@
+import React, { useEffect } from 'react';
+import usePagination from '../../hooks/usePagination';
+import { SimpleStolenBikes } from '../../interfaces/SimpleStolenBikesInterface';
+import Card from '../Card/Card';
+
+interface Props {
+  list: SimpleStolenBikes[]
+}
+export default function DisplayCards({ list }: Props) {
+
+  const {
+    page, getPage, index, prevPage, nextPage, goLast, goFirst, 
+  } = usePagination(list);
+
+  useEffect(() => {
+    getPage();
+  }, [page]);
+
+  return (
+    <div>
+      <span>{index}</span>
+      <button type="button" onClick={prevPage}>back</button>
+      <button type="button" onClick={nextPage}>next</button>
+      <button type="button" onClick={goFirst}>first</button>
+      <button type="button" onClick={goLast}>last</button>
+      
+      {page.map((element, i) => (<Card bike={element} key={i.toString()} />))}
+    </div>
+  );
+}
