+++ stolen-bikes-index-app/src/components/Home/Home.tsx
@@ -1,18 +1,14 @@
 import React, { useEffect } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
-import usePagination from '../../hooks/usePagination';
 import { getStolenBikes } from '../../redux/actions';
 import { State } from '../../redux/reducer';
-import Card from '../Card/Card';
-import HandlePag from '../HandlePag/HandlePag';
+import DisplayCards from '../DisplayCards/DisplayCards';
 import SearchBar from '../SearchBar/SearchBar';
 import './Home.css';
 
 export default function Home() {
   const dispatch = useDispatch();
   const allItems = useSelector((state:State) => state.stolenBikes);
-
-  const { page } = usePagination(allItems);
  
   useEffect(() => {
     if (!allItems.length) { dispatch(getStolenBikes()); }
@@ -20,10 +16,9 @@ export default function Home() {
 
   return (
     <div className="homeCtn">
-      <SearchBar />
-      <HandlePag />
       <h1>Stolen bikes</h1>
-      {page.map((element) => (<Card bike={element} />))}
+      <SearchBar />
+      <DisplayCards list={allItems} />
     </div>
   );
 }
