+++ stolen-bikes-index-app/src/redux/reducer/index.ts
@@ -1,14 +1,12 @@
 import { SimpleStolenBikes } from '../../interfaces/SimpleStolenBikesInterface';
-import { Action, CHANGE_PAGE, GET_STOLEN_BIKES } from '../actions';
+import { Action, GET_STOLEN_BIKES } from '../actions';
 
 export type State = {
   stolenBikes: SimpleStolenBikes[];
-  actualPage: number
 }
 
 const InitialState: State = {
   stolenBikes: [],
-  actualPage: 0,
 };
 
 export default function rootReducer(state:State = InitialState, action:Action):State {
@@ -17,13 +15,8 @@ export default function rootReducer(state:State = InitialState, action:Action):S
       return {
         ...state,
         stolenBikes: action.payload,
-        actualPage: 1,
       };
-    case CHANGE_PAGE:
-      return {
-        ...state,
-        actualPage: action.payload,
-      };
+   
     default: return state;
   }
 }
